locally running vercel integrated with supabase
