+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .env
 db-config.js
 .vercel
+.env*.local
