+++ index.html
@@ -814,7 +814,7 @@
 	<script type="module" src="./db-config.js"></script>
 
 	<!-- other scripts -->
-	<script type="module" src="./api/store-record.js"></script>
+	<script type="module" src="./client.js"></script>
 	<script src="./index.js"></script>
 	<script src="https://js.stripe.com/v3/"></script>
